site tags

+++ click.go
@@ -63,7 +63,7 @@ type ClickID struct {
 	ProfitMargin      int       `json:"profit_margin" msgpack:"profit_margin"`
 	ModelNames        []string  `json:"model_names" msgpack:"-"`
 	Proxy             bool      `json:"proxy" msgpack:"-"`
-	SiteTags          []int64   `json:"site_tags" msgpack:"-"`
+	SiteTags          []int     `json:"site_tags" msgpack:"-"`
 }
 
 // NewClickID ...
